+++ src/extensions/chroma/lightID.ts
@@ -1175,6 +1175,7 @@ export const LightIDList: {
    HipHopEnvironment: {},
    ColliderEnvironment: {},
    BritneyEnvironment: {},
+   Monstercat2Environment: {},
    GlassDesertEnvironment: {
       '0': [1, 2, 3, 4, 5, 6],
       '1': [1, 2, 3, 4, 5, 6, 7, 8],